finsih scan selection and ms level

+++ src/main/java/io/github/mzmine/modules/visualization/rawdataoverviewims/IMSRawDataOverviewPane.java
@@ -410,7 +410,7 @@ public class IMSRawDataOverviewPane extends BorderPane {
     ticChart.cursorPositionProperty().addListener(
         ((observable, oldValue, newValue) -> setSelectedFrame((Frame) newValue.getScan())));
     ticChart.getMouseAdapter().addGestureHandler(new SimpleDataDragGestureHandler((start, end) -> {
-      Range<Float> rtRange = Range.closed((float) start.getX(), (float) end.getX());
+      Range<Double> rtRange = Range.closed(start.getX(), end.getX());
       final ScanSelection selection = scanSelection.cloneWithNewRtRange(rtRange);
       MZmineCore.getTaskController().addTask(
           new MergeFrameThread(rawDataFile, selection, binWidth, mobilityScanNoiseLevel,
